# 49/implement-email-sender/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing Ecommerce",
        "title": "Dashboard Ecommerce Team 2",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/dashboard/best-item-list": {
            "get": {
                "description": "Retrieve a list of the best-selling products",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get best product list",
                "responses": {
                    "200": {
                        "description": "best product list successfully retrieved",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Error getting best product list",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/current-month-earning": {
            "get": {
                "description": "Retrieve the earnings for the current month",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get current month earnings",
                "responses": {
                    "200": {
                        "description": "current month earnings successfully retrieved",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Error getting earnings",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/revenue-chart": {
            "get": {
                "description": "Generate a revenue chart for the dashboard",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get revenue chart",
                "responses": {
                    "200": {
                        "description": "\u003c!DOCTYPE html\u003e\u003chtml\u003e\u003chead\u003e\u003cmeta charset='utf-8'\u003e\u003ctitle\u003eAwesome go-echarts\u003c/title\u003e\u003cscript src='https://go-echarts.github.io/go-echarts-assets/assets/echarts.min.js'\u003e\u003c/script\u003e\u003c/head\u003e\u003cbody\u003e\u003cdiv class='container'\u003e\u003cdiv class='item' id='PvBdsRyYJxot' style='width:900px;height:500px;'\u003e\u003c/div\u003e\u003c/div\u003e\u003cscript type='text/javascript'\u003e\\\"use strict\\\";let goecharts_PvBdsRyYJxot = echarts.init(document.getElementById('PvBdsRyYJxot'), 'white', { renderer: 'canvas' });let option_PvBdsRyYJxot = {\\\"color\\\":[\\\"#5470c6\\\",\\\"#91cc75\\\",\\\"#fac858\\\",\\\"#ee6666\\\",\\\"#73c0de\\\",\\\"#3ba272\\\",\\\"#fc8452\\\",\\\"#9a60b4\\\",\\\"#ea7ccc\\\"],\\\"legend\\\":{},\\\"series\\\":[{\\\"name\\\":\\\"Revenue\\\",\\\"type\\\":\\\"line\\\",\\\"smooth\\\":true,\\\"data\\\":[{\\\"value\\\":150.75},{\\\"value\\\":300},{\\\"value\\\":500.5},{\\\"value\\\":175.25},{\\\"value\\\":250},{\\\"value\\\":100.75},{\\\"value\\\":400.5},{\\\"value\\\":300.25},{\\\"value\\\":275},{\\\"value\\\":125.5},{\\\"value\\\":500},{\\\"value\\\":350}]}],\\\"title\\\":{\\\"text\\\":\\\"Monthly Revenue\\\"},\\\"toolbox\\\":{},\\\"tooltip\\\":{},\\\"xAxis\\\":[{\\\"name\\\":\\\"Month\\\",\\\"data\\\":[\\\"January  \\\",\\\"February \\\",\\\"March    \\\",\\\"April    \\\",\\\"May      \\\",\\\"June     \\\",\\\"July     \\\",\\\"August   \\\",\\\"September\\\",\\\"October  \\\",\\\"November \\\",\\\"December \\\"]}],\\\"yAxis\\\":[{\\\"name\\\":\\\"Revenue\\\"}]}goecharts_PvBdsRyYJxot.setOption(option_PvBdsRyYJxot);\u003c/script\u003e\u003cstyle\u003e.container {margin-top:30px; display: flex;justify-content: center;align-items: center;}.item {margin: auto;}\u003c/style\u003e\u003c/body\u003e\u003c/html\u003e",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error generating revenue chart",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/dashboard/summary": {
            "get": {
                "description": "Retrieve a summary of the dashboard",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get dashboard summary",
                "responses": {
                    "200": {
                        "description": "dashboard summary successfully retrieved",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Error getting summary",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "helper.HTTPResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Authentication": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "UserID": {
            "type": "apiKey",
            "name": "User-ID",
            "in": "header"
        },
        "UserRole": {
            "type": "apiKey",
            "name": "User-Role",
            "in": "header"
        }
    }
}